+++ src/Controller/PlanportadoresAreaController.php
@@ -19,35 +19,37 @@ class PlanportadoresAreaController extends Controller
     /**
      * @Route("/{id}/list", name="planportadoresarea_index", methods="GET",options={"expose"=true})
      */
-    public function index(Request $request,Planportadores $planportadores): Response
+    public function index(Request $request, Planportadores $planportadores): Response
     {
-        if(!$request->isXmlHttpRequest())
+        if (!$request->isXmlHttpRequest())
             throw $this->createAccessDeniedException();
 
-            $planportadoresareas = $this->getDoctrine()->getManager()->createQuery('SELECT ppa.id , a.nombre as area, ppa.categoria, ppa.valor FROM App:PlanportadoresArea ppa JOIN ppa.areas a JOIN ppa.planportadores p WHERE p.id= :id')->setParameter('id',$planportadores->getId())->getResult();
-            return new JsonResponse(
-                $result = [
-                    'iTotalRecords'        => count($planportadoresareas),
-                    'iTotalDisplayRecords' => 10,
-                    'sEcho'                => 0,
-                    'sColumns'             => '',
-                    'aaData'               => $planportadoresareas,
-                ]
-                );
+        $this->denyAccessUnlessGranted('VIEW', $planportadores);
+        $planportadoresareas = $this->getDoctrine()->getManager()->createQuery('SELECT ppa.id , a.nombre as area, ppa.categoria, ppa.valor FROM App:PlanportadoresArea ppa JOIN ppa.areas a JOIN ppa.planportadores p WHERE p.id= :id')->setParameter('id', $planportadores->getId())->getResult();
+        return new JsonResponse(
+            $result = [
+                'iTotalRecords' => count($planportadoresareas),
+                'iTotalDisplayRecords' => 10,
+                'sEcho' => 0,
+                'sColumns' => '',
+                'aaData' => $planportadoresareas,
+            ]
+        );
     }
 
     /**
      * @Route("/{id}/new", name="planportadoresarea_new", methods="GET|POST",options={"expose"=true})
      */
-    public function new(Request $request,Planportadores $planportadores): Response
+    public function new(Request $request, Planportadores $planportadores): Response
     {
-        if(!$request->isXmlHttpRequest())
+        if (!$request->isXmlHttpRequest())
             throw $this->createAccessDeniedException();
-        
+
+        $this->denyAccessUnlessGranted('VIEW', $planportadores);
         $planportadoresarea = new PlanportadoresArea();
         $planportadoresarea->setPlanportadores($planportadores);
         $planportadoresarea->setUsuario($this->getUser());
-        $form = $this->createForm(PlanportadoresAreaType::class, $planportadoresarea, array('institucion'=>$this->getUser()->getInstitucion()->getId(),'action' => $this->generateUrl('planportadoresarea_new',['id'=>$planportadores->getId()])));
+        $form = $this->createForm(PlanportadoresAreaType::class, $planportadoresarea, array('action' => $this->generateUrl('planportadoresarea_new', ['id' => $planportadores->getId()])));
         $form->handleRequest($request);
         $em = $this->getDoctrine()->getManager();
 
@@ -55,8 +57,8 @@ class PlanportadoresAreaController extends Controller
             if ($form->isValid()) {
                 $em->persist($planportadoresarea);
                 $em->flush();
-                $this->addFlash('success',"El plan de portadores fue registrado satisfactoriamente");
-                return new JsonResponse(['estado' =>1]);
+                $this->addFlash('success', "El plan de portadores fue registrado satisfactoriamente");
+                return new JsonResponse(['estado' => 1]);
             } else {
                 $page = $this->renderView('planportadoresarea/_form.html.twig', array(
                     'form' => $form->createView(),
@@ -76,10 +78,11 @@ class PlanportadoresAreaController extends Controller
      */
     public function show(Request $request, PlanportadoresArea $planportadoresarea): Response
     {
-        if(!$request->isXmlHttpRequest())
+        if (!$request->isXmlHttpRequest())
             throw $this->createAccessDeniedException();
-        
-        return $this->render('planportadoresarea/_show.html.twig',['plan'=>$planportadoresarea]);
+
+        $this->denyAccessUnlessGranted('VIEW', $planportadoresarea->getPlanportadores());
+        return $this->render('planportadoresarea/_show.html.twig', ['plan' => $planportadoresarea]);
     }
 
     /**
@@ -87,10 +90,11 @@ class PlanportadoresAreaController extends Controller
      */
     public function edit(Request $request, PlanportadoresArea $planportadoresarea): Response
     {
-        if(!$request->isXmlHttpRequest())
+        if (!$request->isXmlHttpRequest())
             throw $this->createAccessDeniedException();
-        
-        $form = $this->createForm(PlanportadoresAreaType::class, $planportadoresarea, array('institucion'=>$this->getUser()->getInstitucion()->getId(),'action' => $this->generateUrl('planportadoresarea_edit',array('id'=>$planportadoresarea->getId()))));
+
+        $this->denyAccessUnlessGranted('VIEW', $planportadoresarea->getPlanportadores());
+        $form = $this->createForm(PlanportadoresAreaType::class, $planportadoresarea, array('action' => $this->generateUrl('planportadoresarea_edit', array('id' => $planportadoresarea->getId()))));
         $form->handleRequest($request);
         $em = $this->getDoctrine()->getManager();
         if ($form->isSubmitted())
@@ -98,8 +102,8 @@ class PlanportadoresAreaController extends Controller
                 $planportadoresarea->setUsuario($this->getUser());
                 $em->persist($planportadoresarea);
                 $em->flush();
-                $this->addFlash('success',"El plan de portadores fue actualizado satisfactoriamente");
-                return new JsonResponse(['estado' =>1]);
+                $this->addFlash('success', "El plan de portadores fue actualizado satisfactoriamente");
+                return new JsonResponse(['estado' => 1]);
             } else {
                 $page = $this->renderView('planportadoresarea/_form.html.twig', array(
                     'form' => $form->createView(),
@@ -126,10 +130,11 @@ class PlanportadoresAreaController extends Controller
         if (!$request->isXmlHttpRequest())
             throw $this->createAccessDeniedException();
 
+        $this->denyAccessUnlessGranted('VIEW', $planportadoresarea->getPlanportadores());
         $em = $this->getDoctrine()->getManager();
         $em->remove($planportadoresarea);
         $em->flush();
-        $this->addFlash('success',"El plan de portadores fue eliminado satisfactoriamente");
-        return new JsonResponse(['estado' =>1]);
+        $this->addFlash('success', "El plan de portadores fue eliminado satisfactoriamente");
+        return new JsonResponse(['estado' => 1]);
     }
 }
